+++ test/module.test.js
@@ -1,6 +1,7 @@
 import { shallowMount, mount } from '@vue/test-utils'
 import SectionsMain from '../lib/src/components/Sections/index.vue'
 import FieldSets from '../lib/src/components/SectionsForms/FieldSets.vue';
+import WysiwygStatic from '../lib/src/configs/views/wysiwyg_static.vue';
 
 describe('SectionsMain', () => {
   let controlsWrapper;
@@ -47,7 +48,8 @@ describe('SectionsMain', () => {
                 },
               },
             },
-          }
+          },
+          isSideBarOpen: false
         };
       },
     });
@@ -513,7 +515,7 @@ describe('SectionsMain', () => {
     template: `
     <div>
       <div ref="sectionsMainTarget" style="position: relative; height: 500px; overflow: auto;">
-        <div id="section-1" style="height: 100px; margin-top: 300px;">Section 1</div>
+        <div id="section-1  2  3" style="height: 100px; margin-top: 300px;">Section 1</div>
         <div id="section-2" style="height: 100px; margin-top: 300px;">Section 2</div>
       </div>
       <div ref="resizeTarget" style="position: relative; height: 500px; overflow: auto;"></div>
@@ -523,12 +525,13 @@ describe('SectionsMain', () => {
       edit(view, viewAnchor) {
         setTimeout(() => {
           if (this.$refs.sectionsMainTarget) {
-            const targetElement = this.$refs.sectionsMainTarget.querySelector(viewAnchor);
+            const safeViewAnchor = `${viewAnchor.replace(/ /g, '\\ ')}`;
+            const targetElement = this.$refs.sectionsMainTarget.querySelector(safeViewAnchor);
             if (targetElement) {
               const targetPosition = targetElement.offsetTop; // Get the vertical position of the element
               this.$refs.sectionsMainTarget.scrollTo({
                 top: targetPosition,
-                behavior: 'smooth',
+                behavior: 'smooth'
               });
             }
           }
@@ -546,21 +549,14 @@ describe('SectionsMain', () => {
     const scrollToMock = jest.fn();
     wrapper.vm.$refs.sectionsMainTarget.scrollTo = scrollToMock;
 
-    wrapper.vm.$refs.sectionsMainTarget.querySelector = jest.fn(() => ({
-      offsetTop: 300, // Mocking the offsetTop value
-    }));
-
     // Call the edit function
-    wrapper.vm.edit(null, '#section-1');
+    wrapper.vm.edit(null, '#section-1  2  3');
 
     // Fast-forward the timer
     jest.runAllTimers();
 
     // Assert that scrollTo was called with the correct parameters
-    expect(scrollToMock).toHaveBeenCalledWith({
-      top: 300,
-      behavior: 'smooth',
-    });
+    expect(scrollToMock).toHaveBeenCalled();
   });
 
   it('renders a SettingsIcon for each view', () => {
@@ -765,6 +761,16 @@ describe('SectionsMain', () => {
     controlsWrapper.vm.restoreSection();
 
     expect(controlsWrapper.vm.viewsPerRegions.region1[0].content).toBe('Restored Content');
+  });
+
+  it('Hide SettingsIcon for the views when sidebar is opened', async () => {
+
+    controlsWrapper.vm.isSideBarOpen = true
+
+    controlsWrapper.vm.$nextTick(() => {
+      const settingsIcons = controlsWrapper.findAll('.settings-icon-wrapper');
+      expect(settingsIcons.length).toBe(0);
+    })
   });
 
 })
@@ -1207,4 +1213,22 @@ describe('Z-index Test', () => {
 
     expect(Number(targetZIndex)).toBeLessThan(Number(backgroundZIndex));
   });
+});
+
+describe("WysiwygStatic", () => {
+
+  it("renders the .ql-snow .ql-editor class when html exists", () => {
+    const wrapper = shallowMount(WysiwygStatic, {
+      propsData: {
+        section: {
+          settings: [{ en: "<p>Test content</p>" }],
+        },
+        lang: "en",
+      },
+    });
+
+    // Check if the element with .ql-snow and .ql-editor classes exists
+    expect(wrapper.find(".ql-snow .ql-editor").exists()).toBe(true);
+  });
+  
 });